+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 # Changelog
 
+### 1.0.1 
+* Fixed bug with IPv6 where : was a disallowed character in the cache key.
+* Changed the HTML "step" from 60 to 1 – I had thought it would step by 60 but still allow % values.
+
+